requirements.txt
# sudo apt-get install wkhtmltopdf
pdfkit
requests
jinja
pycups
